requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
idna==2.10
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.3
